% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{Daisy}
\alias{Daisy}
\title{Running Dasiy App. to estimate and forecast the soil water content and matric water potential}
\usage{
Daisy(WeatherData)
}
\arguments{
\item{WeatherData}{The weather information}
}
\value{
Estimation and forecasting of soil water content(kg/kg) and matric potential water(log(h))
}
\description{
Running Dasiy App. to estimate and forecast the soil water content and matric water potential
}
