% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{DLMfilter}
\alias{DLMfilter}
\title{Gaussian SSM filtering}
\usage{
DLMfilter(mod, Mean_DaisyOutputs, W, V)
}
\arguments{
\item{mod}{set of parameters needed in the GSSM model.}

\item{Mean_DaisyOutputs}{A list contains of Means of soil water content.}

\item{W}{System variance of the GSSM.}

\item{V}{Observation variance of the GSSM.}
}
\value{
Updated information of filtering.
}
\description{
Gaussian SSM filtering
}
