% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{DWF_File}
\alias{DWF_File}
\title{Creating a .dwf file to run Dasiy}
\usage{
DWF_File(WeatherData)
}
\arguments{
\item{WeatherData}{The weather information as inputs for building a .dwf file to run Daisy App.}
}
\value{
A .dwf file as an input for Daisy
}
\description{
Creating a .dwf file to run Dasiy
}
