% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{EM}
\alias{EM}
\title{EM algorithm for estimation of system variance in Gaussian SSM}
\usage{
EM(mod, fdlm1, sdlm1, Wm)
}
\arguments{
\item{mod}{Set of parameters needed in the GSSM model.}

\item{fdlm1}{Output information of filtering the GSSM using the function DLMfilter.}

\item{sdlm1}{Output information of smoothing the GSSM using the function Smoother.}

\item{Wm}{Updated system variance of the GSSM.}
}
\value{
Updated information of system variance W
}
\description{
EM algorithm for estimation of system variance in Gaussian SSM
}
