% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{package}
\name{FieldReadiness}
\alias{FieldReadiness}
\title{Optimization model: MDP Statistical model: Gaussian state space model.}
\description{
This package includes codes for optimization and statistical models
To find an optimal schedule plan for tillage operations.
}
\section{History}{

\itemize{
\item 28-09-2016: Start to Build the package
}
}

\section{Question/Do}{

\itemize{
\item ?
}
}

\author{
Parisa Bagheri Tookanlou \email{pbagheri@eng.au.dk}.
}
