% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{findIndex}
\alias{findIndex}
\title{Function to find the index of a state given state value}
\usage{
findIndex(st, dis)
}
\arguments{
\item{st}{State value}

\item{dis}{A matrix containing the discritization values of a state variable}
}
\value{
An index related to the value of st
}
\description{
Function to find the index of a state given state value
}
