% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{Mean_Daisy}
\alias{Mean_Daisy}
\title{Means of Daisy outputs}
\usage{
Mean_Daisy(DaisyOutputs)
}
\arguments{
\item{DaisyOutputs}{Obtained outputs from Daisy model}
}
\value{
Means of soil water content and matric water potential
}
\description{
Means of Daisy outputs
}
