% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optActions.R
\name{optimalSearch}
\alias{optimalSearch}
\title{A function for determining the optimal actions and updated information obtained by Gaussian SSM}
\usage{
optimalSearch(
  t_count,
  param,
  policy,
  WatObs,
  temData,
  precData,
  GlobRdData,
  iniTrueWat,
  givenWatInfo
)
}
\arguments{
\item{t_count}{The time period of observing weather data}

\item{param}{Parameter values given in R function \code{setParam}}

\item{policy}{The optimal policy of the MDP}

\item{WatObs}{Given soil water content data (if givenWatInfo==TRUE)}

\item{temData}{Temperature data}

\item{precData}{Precipitation data}

\item{GlobRdData}{Global Radiation data}

\item{iniTrueWat}{Initial soil-water content at day t=1}

\item{givenWatInfo}{A boolean value}
}
\value{
A data frame containing the optimal actions and updated information obtained by Gaussian SSM
}
\description{
A function for determining the optimal actions and updated information obtained by Gaussian SSM
}
