% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{setModel}
\alias{setModel}
\title{Set the parameters of the Gaussian SSM.}
\usage{
setModel(param, t, GG, V, W, m0, C0)
}
\arguments{
\item{param}{Parameter values given in R function \code{setParam}}

\item{t}{Time period of observing sensor data}

\item{GG}{design matrix of observation equation.}

\item{V}{Observation variance of the nGSSM}

\item{W}{System variance of the nGSSM (optional values)}

\item{m0}{Initial mean of posterior distribution at the insertion time (t=0)}

\item{C0}{Initial variance of posterior distribution at the insertion time (t=0)}
}
\value{
A list containing the parameters of the GSSM.
}
\description{
Set the parameters of the Gaussian SSM.
}
