% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramModels.R
\name{setParam}
\alias{setParam}
\title{Set the parameters used when build the MDP model}
\usage{
setParam(
  opNum = 2,
  tMax = 14,
  StartDay = 13,
  LastDay = 30,
  opSeq = c(1, 2),
  opD = c(6, 4),
  opE = c(1, 1 + opD[1]),
  opL = c(tMax - opD[2], tMax),
  opDelay = c(0, 0, 0),
  opFixCost = c(1000, 1000),
  watTh = c(50, 50),
  coefLoss = 0.3,
  priceYield = 10,
  machCap = 50,
  yieldHa = 100,
  fieldArea = 100,
  coefTimeliness = 0.01,
  costSkip = 80000,
  watUpper = c(37, 37),
  watLower = c(28, 28),
  weightCompletion = 1,
  weightWorkable = 1,
  weightTraffic = 1,
  minOpt = 18,
  maxOpt = 23,
  BetaM1 = 0.7,
  Lambda = 2.465,
  Gama = 0.667,
  c_Pos = c(6.108, 0.4437, 0.01429, 0.0002651, 3.031e-06, 2.034e-08, 6.137e-11),
  c_Neg = c(6.109, 0.5035, 0.01886, 0.0004176, 5.825e-06, 4.839e-08, 1.839e-10),
  centerPointsAvgWat = c(20.94923, 23.82599, 26.70275, 29.57951, 32.45627),
  centerPointsAvgMat = c(34.4775, 97.7237, 160.9699, 224.2161, 287.4623),
  centerPointsSdWat = seq(1, 1, by = 1),
  centerPointsSdPos = c(0.005, 0.03, 0.1),
  centerPointsMeanPos = seq(0.8, 1.2, by = 0.15),
  centerPointsTem = c(10.723224, 14.70236, 18.681497),
  centerPointsPre = c(4.55625, 10.63125, 16.70625, 22.78125),
  centerPointsGlobRd = c(69.98952, 134.13013, 198.27074),
  centerPointsEvap = c(11.182115, 23.541657, 37.661322),
  temMeanDry = 14.34038,
  temMeanWet = 14.20856,
  temVarDry = 2.15906,
  temVarWet = 4.20047,
  GRdMeanDry = 165.5937,
  GRdMeanWet = 100.3606,
  GRdVarDry = 749.9474,
  GRdVarWet = 1836.803,
  EvapMeanDry = 2.498701,
  EvapMeanWet = 1.495789,
  EvapVarDry = 0.1684589,
  EvapVarWet = 0.4023626,
  dryDayTh = 0.25,
  precShape = 0.2872144,
  precScale = 7.689959,
  prDryWet = 0.3880316,
  prWetWet = 0.7106343,
  hydroWatR = 0.01,
  hydroWatS = 0.439,
  hydroM = 15,
  hydroKs = 12.061,
  hydroLamba = 0.085 * log(hydroKs) + 0.1574,
  hydroFi = 54.727 * log(hydroKs) - 323.9,
  hydroETa = -2,
  hydroETb = 1.26,
  hydroETx = 0.25,
  hydroVanAlpha = 0.0314,
  hydroVanM = 0.1528,
  hydroVanN = 1.1804,
  hydroBulkDensity = 1.7,
  gSSMW = 0.063,
  gSSMV = 9,
  gSSMm0 = 1,
  gSSMc0 = 0.001,
  nGSSMm0 = 4,
  nGSSMc0 = 2,
  nGSSMK = 20,
  rewRisk = TRUE,
  check = FALSE
)
}
\arguments{
\item{opNum}{Number of operations for scheduling}

\item{tMax}{Maximum length of growing cycle}

\item{StartDay}{The first day of growing cycle}

\item{LastDay}{The last day of growing cycle}

\item{opSeq}{A vector showing operations indexes and sequence}

\item{opD}{A vector showing the time needed to complete operation (obtained by machine capacity and the field area)}

\item{opE}{A vector showing earliest start time of operations}

\item{opL}{A vector showing latest start time of operationss}

\item{opDelay}{A vector including the delay times after finishing  the operations (except the last operation)}

\item{opFixCost}{A vector including the fixed costs of tillage operations}

\item{watTh}{A vector including threshold values used as the optimal level of soil-water content for performing tillage operations}

\item{coefLoss}{Coeficiant showing yield reduction when soil-water content is not appropriate}

\item{priceYield}{Price per kg yield (DDK)}

\item{machCap}{A vector showing the machine capacity for tillage operations}

\item{yieldHa}{Estimated yield per hetar in field (kg)}

\item{fieldArea}{Area of field (ha)}

\item{coefTimeliness}{Timeliness Coeficiant showing yield reduction resulting from postponing a tillage operation.}

\item{costSkip}{Cost of skipping tillage operations for the current cropping period}

\item{watUpper}{Upper limits of water content for workability criterion calculated based on the method in \url{http://www.sciencedirect.com/science/article/pii/S0167198700001549}}

\item{watLower}{Lower limits of water content for workability criterion calculated based on the method in \url{http://www.sciencedirect.com/science/article/pii/S0167198700001549}}

\item{weightCompletion}{Weight of the completion criterion used in the reward function of the MDP.}

\item{weightWorkable}{Weight of the workability criterion used in the reward function of the MDP.}

\item{weightTraffic}{Weight of the trafficability criterion used in the reward function of the MDP.}

\item{minOpt}{The lower tail of the interval related to the best time period for finishing tillage operations.}

\item{maxOpt}{The upper tail of the interval related to the best time period for finishing tillage operations.}

\item{BetaM1}{Parameter used to calculate evatranspiration.}

\item{Lambda}{Parameter used to calculate evatranspiration.}

\item{Gama}{Parameter used to calculate evatranspiration.}

\item{c_Pos}{Coefficient to determine evatranspiration.}

\item{c_Neg}{Coefficient to determine evatranspiration.}

\item{centerPointsAvgWat}{Center points for discritization of estimated mean of soil-water content using historical weather data.}

\item{centerPointsAvgMat}{Center points for discritization of estimated mean of matric water potential using historical weather data.}

\item{centerPointsSdWat}{Center points for discritization of estimated standard deviation of soil-water content}

\item{centerPointsSdPos}{Center points for discritization of estimated standard deviation of posterior distribution in SSM}

\item{centerPointsMeanPos}{Center points for discritization of estimated mean of posterior distribution in SSM}

\item{centerPointsTem}{Center points for discritization of weather forecast data regarding air temperature}

\item{centerPointsPre}{Center points for discritization of weather forecast data regarding precipitation}

\item{centerPointsGlobRd}{Center points for discritization of weather forecast data regarding global radiation}

\item{centerPointsEvap}{Center points for discritization of weather forecast data regarding evatranspiration}

\item{temMeanDry}{Mean of air temprature in dry days (\url{http://link.springer.com/article/10.1007/BF00142466}).}

\item{temMeanWet}{Mean of air temprature in wet days (\url{http://link.springer.com/article/10.1007/BF00142466}).}

\item{temVarDry}{Variance of air temprature in dry days (\url{http://link.springer.com/article/10.1007/BF00142466}).}

\item{temVarWet}{Variance of air temprature in wet days (\url{http://link.springer.com/article/10.1007/BF00142466}).}

\item{GRdMeanDry}{Mean of global radiation in dry days}

\item{GRdMeanWet}{Mean of global radiation in wet days}

\item{GRdVarDry}{Variance of global radiation in dry days}

\item{GRdVarWet}{Variance of global radiation in wet days}

\item{EvapMeanDry}{Mean of evatranspiration in dry days}

\item{EvapMeanWet}{Mean of evatranspiration in wet days}

\item{EvapVarDry}{Variance of evatranspiration in dry days}

\item{EvapVarWet}{Variance of evatranspiration in wet days}

\item{dryDayTh}{Threshold of precipitation amount for being a dry day (\url{http://link.springer.com/article/10.1007/BF00142466}).}

\item{precShape}{Shape parameter of gamma distribution for precipitation amount (\url{http://link.springer.com/article/10.1007/BF00142466}).}

\item{precScale}{Scale parameter of gamma distribution for precipitation amount (\url{http://link.springer.com/article/10.1007/BF00142466}).}

\item{prDryWet}{Probability of a wet day when the previous day is dry (\url{http://link.springer.com/article/10.1007/BF00142466}).}

\item{prWetWet}{Probability of a wet day when the previous day is wet (\url{http://link.springer.com/article/10.1007/BF00142466}).}

\item{hydroWatR}{Residual soil moisture value (\url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.6629/abstract}).}

\item{hydroWatS}{Soil moisture value at saturation (\url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.6629/abstract}).}

\item{hydroM}{Parameter used in the infiltration process of rainfall-runoff model (\url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.6629/abstract}).}

\item{hydroKs}{Saturated hydralic conductivity used for drainage process in the rainfall-runoff model (\url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.6629/abstract}).}

\item{hydroLamba}{Parameter used for drainage process in therainfall-runoff model (\url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.6629/abstract}).}

\item{hydroFi}{Parameter used in the infiltration process of rainfall-runoff model (\url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.6629/abstract}).}

\item{hydroETa}{Parameter used for evapotranpiration process in the rainfall-runoff model (\url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.6629/abstract}).}

\item{hydroETb}{Parameter used for evapotranpiration process in the rainfall-runoff model (\url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.6629/abstract}).}

\item{hydroETx}{Parameter used for evapotranpiration process in the rainfall-runoff model (\url{http://onlinelibrary.wiley.com/doi/10.1002/hyp.6629/abstract}).}

\item{hydroVanAlpha}{Parameter used in van Genuchten equation to calculate the matric water potensial (\url{http://www.sciencedirect.com/science/article/pii/S0016706198001323})}

\item{hydroVanM}{Parameter used in van Genuchten equation to calculate the matric water potensial (\url{http://www.sciencedirect.com/science/article/pii/S0016706198001323})}

\item{hydroVanN}{Parameter used in van Genuchten equation to calculate the matric water potensial (\url{http://www.sciencedirect.com/science/article/pii/S0016706198001323})}

\item{hydroBulkDensity}{Parameter used to change the unit of soil-water content from kg/kg to cm^3/cm^3(\url{http://www.sciencedirect.com/science/article/pii/S0016706198001323})}

\item{gSSMW}{System variance of Gaussian SSM.}

\item{gSSMV}{Observation variance of Gaussian SSM.}

\item{gSSMm0}{Initial posterior mean of Gaussian SSM.}

\item{gSSMc0}{Initial posterior variance of Gaussian SSM.}

\item{nGSSMm0}{Initial posterior mean of non-Gaussian SSM.}

\item{nGSSMc0}{Initial posterior mean of non-Gaussian SSM.}

\item{nGSSMK}{Number of observations in non-Gaussian SSM.}

\item{rewRisk}{A boolean variable specifing to calculate the reward based on cost parameters or the satisfaction level for trafficability, workability and completion criteria.}

\item{check}{Check model e.g. do trans pr sum to one.}
}
\value{
A list containing all the parameters used in three-level MDP
}
\description{
Set the parameters used when build the MDP model
}
