% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{simMois}
\alias{simMois}
\title{Simulate the observed and true soil water content given weather data and initial soil-water condition (at time t=1)}
\usage{
simMois(t_count, temData, precData, GlobRdData, iniTrueWat, param)
}
\arguments{
\item{t_count}{The time period of observing weather data}

\item{temData}{Weather data regarding temperature}

\item{precData}{Weather data regarding precipitation}

\item{GlobRdData}{Weather data regarding global radiation}

\item{iniTrueWat}{An initial value for soil water content}

\item{param}{Parameter values given in R function \code{setParam}}
}
\value{
A list containing simulated observed and true soil water content.
}
\description{
Simulate the observed and true soil water content given weather data and initial soil-water condition (at time t=1)
}
