% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{Smoother}
\alias{Smoother}
\title{Gaussian SSM smoothing}
\usage{
Smoother(mod, fdlm1, W)
}
\arguments{
\item{mod}{set of parameters needed in the GSSM model.}

\item{fdlm1}{output information of filtering the GSSM using the function DLMfilter.}

\item{W}{System variance of the GSSM.}
}
\value{
Updated information of smoothing
}
\description{
Gaussian SSM smoothing
}
