% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SolveMDPModel}
\alias{SolveMDPModel}
\title{Solving the MDP using value iteration algorithm.}
\usage{
SolveMDPModel(paramModel)
}
\arguments{
\item{paramModel}{parameters a list created using \code{\link{setParam}}.}
}
\value{
A list
}
\description{
Solving the MDP using value iteration algorithm.
}
