% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{Terranimo_SCI}
\alias{Terranimo_SCI}
\title{Running Terranimo.world App. to calculate Soil Compaction Index(SCI)}
\usage{
Terranimo_SCI(disAvgMat)
}
\arguments{
\item{disAvgMat}{Soil water content}
}
\value{
Soil compaction index of the soil at time t for different layers
}
\description{
Running Terranimo.world App. to calculate Soil Compaction Index(SCI)
}
