% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{VanGe}
\alias{VanGe}
\title{van Genuchten equation to convert matric water potential to gravimetric soil water content
Van Genuchten equation can be found in \url{http://www.sciencedirect.com/science/article/pii/S0167198700001549}.}
\usage{
VanGe(h)
}
\arguments{
\item{h}{Matric water potential in hPa unit}
}
\value{
Gravimetric soil water content in cm3/cm3 unit
}
\description{
van Genuchten equation to convert matric water potential to gravimetric soil water content
Van Genuchten equation can be found in \url{http://www.sciencedirect.com/science/article/pii/S0167198700001549}.
}
