% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statModels.R
\name{WeatherInfo}
\alias{WeatherInfo}
\title{Creating a list of historical and forecasting weather Data as an input argument for running Daisy App.}
\usage{
WeatherInfo(t_count, param, HistMetData1, HistMetData2, PredMetData)
}
\arguments{
\item{t_count}{The time period of observing weather data}

\item{param}{Parameter values given in R function \code{setParam}}

\item{HistMetData1}{The hiatorical weather data}

\item{HistMetData2}{The historical weather data during growing cycle}

\item{PredMetData}{The prediction or simulater weather data during growing cycle}
}
\value{
A list containing weather information we need to run Daisy
}
\description{
Creating a list of historical and forecasting weather Data as an input argument for running Daisy App.
}
